requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==0.8.11
Flask-Login==0.6.3
PyMongo==4.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
reportlab==4.0.4
